Rename comment/id router to comment/commentid
명확한 이름 구별을 위해 comment/id 를 comment/commentid 로 변경
