+++ src/routes/study/comment/commentid/index.ts
